more comments, rename SerializedPhotoLevel to something more reusable (for challenges)
