+++ Refresh.GameServer/Types/Challenges/LbpHub/SerializedChallenge.cs
@@ -16,7 +16,7 @@ public class SerializedChallenge : IDataConvertableFrom<SerializedChallenge, Gam
     /// <summary>
     /// This challenge's level's type (developer/user) and id (or story id if it is a developer level).
     /// </summary>
-    [XmlElement("slot")] public SerializedPhotoLevel Level { get; set; }
+    [XmlElement("slot")] public SerializedLevelIdTypeName Level { get; set; }
     [XmlElement("author")] public string PublisherName { get; set; } = SystemUsers.UnknownUserName;
     /// <summary>
     /// Always 0 when challenge is first uploaded by LBP hub, doesn't appear to affect anything if set to not 0 in the response.
@@ -59,7 +59,7 @@ public class SerializedChallenge : IDataConvertableFrom<SerializedChallenge, Gam
         {
             ChallengeId = old.ChallengeId,
             Name = old.Name,
-            Level = new SerializedPhotoLevel
+            Level = new SerializedLevelIdTypeName
             {
                 LevelId = old.Level.SlotType == GameSlotType.Story ? old.Level.StoryId : old.Level.LevelId,
                 Type = old.Level.SlotType.ToGameType(),
