+++ Refresh.GameServer/Types/Levels/SerializedLevelIdTypeName.cs
@@ -1,12 +1,12 @@
 using System.Xml.Serialization;
 
-namespace Refresh.GameServer.Types.Photos;
+namespace Refresh.GameServer.Types.Levels;
 
 #nullable disable
 
 [XmlRoot("slot")]
 [XmlType("slot")]
-public class SerializedPhotoLevel
+public class SerializedLevelIdTypeName // I ran out of creative class names
 {
     [XmlElement("id")] public int LevelId { get; set; }
     [XmlElement("name")] public string Title { get; set; }
