+++ Refresh.GameServer/Types/Photos/SerializedPhoto.cs
@@ -1,8 +1,6 @@
 using System.Xml.Serialization;
-using Bunkum.Core.Storage;
-using Refresh.GameServer.Authentication;
-using Refresh.GameServer.Database;
 using Refresh.GameServer.Types.Data;
+using Refresh.GameServer.Types.Levels;
 
 namespace Refresh.GameServer.Types.Photos;
 
@@ -26,7 +24,7 @@ public class SerializedPhoto
     [XmlElement("large")] public string LargeHash { get; set; }
     [XmlElement("plan")] public string PlanHash { get; set; }
     
-    [XmlElement("slot")] public SerializedPhotoLevel Level { get; set; }
+    [XmlElement("slot")] public SerializedLevelIdTypeName Level { get; set; }
     
     [XmlArray("subjects")] public List<SerializedPhotoSubject> PhotoSubjects { get; set; }
